Assorted optimizations
In my benchmarks, compression is ~25% faster and decompression is ~50%
faster.

+++ trie/trie.go
@@ -1,15 +1,15 @@
-// Package trie is an implementation of a trie (prefix tree) data structure mapping []bytes to ints. It
-// provides a small and simple API for usage as a set as well as a 'Node' API for walking the trie.
+// Package trie is an implementation of a trie (prefix tree) data structure
+// mapping []bytes to bytes. It provides a small and simple API for usage as a
+// set as well as a 'Node' API for walking the trie.
 package trie
 
 // A Node represents a logical vertex in the trie structure.
 type Node struct {
 	branches    [256]*Node
-	val int
+	val         byte
 	terminal    bool
 	hasChildren bool
 }
-
 
 // A Trie is a a prefix tree.
 type Trie struct {
@@ -23,9 +23,9 @@ func New() *Trie {
 	}
 }
 
-// Put inserts the mapping k -> v into the Trie, overwriting any previous value. It returns true if the
-// element was not previously in t.
-func (t *Trie) Put(k []byte, v int) bool {
+// Put inserts the mapping k -> v into the Trie, overwriting any previous value.
+// It returns true if the element was not previously in t.
+func (t *Trie) Put(k []byte, v byte) bool {
 	n := t.root
 	for _, c := range k {
 		next, ok := n.Walk(c)
@@ -45,7 +45,7 @@ func (t *Trie) Put(k []byte, v int) bool {
 }
 
 // Get the value corresponding to k in t, if any.
-func (t *Trie) Get(k []byte) (v int, ok bool) {
+func (t *Trie) Get(k []byte) (v byte, ok bool) {
 	n := t.root
 	for _, c := range k {
 		next, ok := n.Walk(c)
@@ -60,29 +60,31 @@ func (t *Trie) Get(k []byte) (v int, ok bool) {
 	return 0, false
 }
 
-// Root returns the root node of a Trie. A valid Trie (i.e., constructed with New), always has a non-nil root
-// node.
+// Root returns the root node of a Trie.
+// A valid Trie (i.e., constructed with New), always has a non-nil root node.
 func (t *Trie) Root() *Node { return t.root }
 
-// Walk returns the node reached along edge c, if one exists. The ok value indicates whether such a node
-// exists.
+// Walk returns the node reached along edge c, if one exists.
+// The ok value indicates whether such a node exists.
 func (n *Node) Walk(c byte) (next *Node, ok bool) {
 	next = n.branches[int(c)]
 	return next, (next != nil)
 }
 
-// Terminal indicates whether n is terminal in the trie (that is, whether the path from the root to n
-// represents an element in the set). For instance, if the root node is terminal, then []byte{} is in the
-// trie.
+// Terminal indicates whether n is terminal in the trie (that is, whether the
+// path from the root to n represents an element in the set).
+// For instance, if/ the root node is terminal, then []byte{} is in the trie.
 func (n *Node) Terminal() bool { return n.terminal }
 
-// Leaf indicates whether n is a leaf node in the trie (that is, whether it has children). A leaf node must be
-// terminal (else it would not exist). Logically, if n is a leaf node then the []byte represented by the path
-// from the root to n is not a proper prefix of any element of the trie.
+// Leaf indicates whether n is a leaf node in the trie (that is, whether it has
+// children). A leaf node must be terminal (else it would not exist).
+// Logically, if n is a leaf node then the []byte represented by the path from
+// the root to n is not a proper prefix of any element of the trie.
 func (n *Node) Leaf() bool { return !n.hasChildren }
 
-// Val gives the value associated with this node. It panics if n is not terminal.
-func (n *Node) Val() int {
+// Val gives the value associated with this node.
+// It panics if n is not terminal.
+func (n *Node) Val() byte {
 	if !n.terminal {
 		panic("Val called on non-terminal node")
 	}
